Address @kmws's review comments

+++ test/test-helpers.cpp
@@ -16,7 +16,7 @@ TEST_CASE("EnvVar object restores the environment variable to its original "
 
 	const auto overwrite = true;
 
-	::setenv(var, expected.c_str(), (int)overwrite);
+	::setenv(var, expected.c_str(), overwrite);
 	REQUIRE(expected == ::getenv(var));
 
 	{
@@ -24,7 +24,7 @@ TEST_CASE("EnvVar object restores the environment variable to its original "
 		REQUIRE(expected == ::getenv(var));
 
 		const auto newValue = std::string("totally new value");
-		::setenv(var, newValue.c_str(), (int)overwrite);
+		::setenv(var, newValue.c_str(), overwrite);
 
 		REQUIRE_FALSE(expected == ::getenv(var));
 	}
@@ -67,7 +67,7 @@ TEST_CASE("EnvVar::set() doesn't change the value to which the environment "
 	const auto expected = std::string("let's try this out, shall we?");
 
 	const auto overwrite = true;
-	::setenv(var, expected.c_str(), (int)overwrite);
+	::setenv(var, expected.c_str(), overwrite);
 	REQUIRE(expected == ::getenv(var));
 
 	{
@@ -93,7 +93,7 @@ TEST_CASE("EnvVar::set() runs a function (set by on_change()) after changing "
 	const auto expected = std::string("let's try this out, shall we?");
 
 	const auto overwrite = true;
-	::setenv(var, expected.c_str(), (int)overwrite);
+	::setenv(var, expected.c_str(), overwrite);
 
 	SECTION("The function is ran *after* the change") {
 		TestHelpers::EnvVar envVar(var);
@@ -141,7 +141,7 @@ TEST_CASE("EnvVar::unset() completely removes the variable from the environment"
 	const auto expected = std::string("let's try this out, shall we?");
 
 	const auto overwrite = true;
-	::setenv(var, expected.c_str(), (int)overwrite);
+	::setenv(var, expected.c_str(), overwrite);
 
 	char* value = ::getenv(var);
 	REQUIRE_FALSE(value == nullptr);
@@ -179,7 +179,7 @@ TEST_CASE("EnvVar::unset() doesn't change the value to which the environment "
 	const auto expected = std::string("let's try this out, shall we?");
 
 	const auto overwrite = true;
-	::setenv(var, expected.c_str(), (int)overwrite);
+	::setenv(var, expected.c_str(), overwrite);
 
 	char* value = ::getenv(var);
 	REQUIRE_FALSE(value == nullptr);
@@ -222,7 +222,7 @@ TEST_CASE("EnvVar::unset() runs a function (set by on_change()) after changing "
 	const auto expected = std::string("let's try this out, shall we?");
 
 	const auto overwrite = true;
-	::setenv(var, expected.c_str(), (int)overwrite);
+	::setenv(var, expected.c_str(), overwrite);
 
 	char* value = ::getenv(var);
 	REQUIRE_FALSE(value == nullptr);
@@ -266,11 +266,20 @@ TEST_CASE("EnvVar's destructor runs a function (set by on_change()) after "
 	const auto expected = std::string("let's try this out, shall we?");
 
 	SECTION("Variable wasn't even set") {
+		// Intentionally left empty.
+		//
+		// Catch framework runs each test case multiple times, each time
+		// entering some section it hasn't entered before. In our situation
+		// here, the test case will be ran twice: once with setenv() being
+		// called, and once without. As a result, EnvVar below will be created
+		// in two different situations: once when the environment variable is
+		// already present, and once when it's absent. The point of the test is
+		// that this wouldn't matter: EnvVar will behave the same regardless.
 	}
 
 	SECTION("Variable was set before EnvVar is created") {
 		const auto overwrite = true;
-		::setenv(var, expected.c_str(), (int)overwrite);
+		::setenv(var, expected.c_str(), overwrite);
 	}
 
 	auto counter = unsigned{};
